requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Computer Vision & Object Detection
ultralytics>=8.0.0  # YOLOv8/v9
mmdet>=3.0.0  # MMDetection framework
albumentations>=1.3.0
opencv-python>=4.8.0
timm>=0.9.0  # PyTorch Image Models

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# Metrics & Evaluation
pycocotools>=2.0.6
torchmetrics>=1.0.0

# Configuration & Utils
pyyaml>=6.0
tqdm>=4.65.0
hydra-core>=1.3.0

# SAHI for small object detection
sahi>=0.11.0

# Optional: Transformer-based detection
transformers>=4.30.0
